revise lr and mom scheduling
